Cast to quell lost precision warning from gcc x86_64

+++ tests/exit2.c
@@ -43,7 +43,7 @@
 void *
 func(void * arg)
 {
-  int failed = (int) arg;
+  int failed = (int)(size_t) arg;
 
   pthread_exit(arg);
 
@@ -51,7 +51,7 @@ func(void * arg)
   /*
    * Trick gcc compiler into not issuing a warning here
    */
-  assert(failed - (int)arg);
+  assert(failed - (int)(size_t)arg);
 
   return NULL;
 }
